More focused rename
